chore: console error when user is null is some cases

+++ src/app/modules/common-components/login-popup/login-popup.component.ts
@@ -131,7 +131,7 @@ export class LoginPopupComponent {
     private oauth(provider: AuthProvider): void {
         const prevUser = this.af.auth.currentUser;
         this.listService.getUserLists(prevUser.uid).subscribe(lists => {
-            if (this.af.auth.currentUser.isAnonymous) {
+            if (this.af.auth.currentUser !== null && this.af.auth.currentUser.isAnonymous) {
                 this.userService.deleteUser(prevUser.uid);
                 this.af.auth.currentUser.delete();
             }
